Update file_revision.rb
Will now check to see if the report returned is empty or nil and stop the processing of that file with an error.
It also now uses fetch to obtain values as you can then submit a default value. Ideally the nil and empty check will stop this from ever being used. 

Also prints '.' on screen to show that something is happening.